Avanza #28: cambia tipo de dato de appliance.go
Cambia el tipo de dato de consumption y aclara el uso de duration como
minutos

+++ internal/appliance.go
@@ -2,11 +2,11 @@ package internal
 
 type Appliance struct {
 	Name string
-	Consumption int
-	Duration int
+	Consumption float32 // expressed in Wh
+	Duration int // expressed in minutes
 }
 
-func NewAppliance(name string, consumption int, duration int) Appliance {
+func NewAppliance(name string, consumption float32, duration int) Appliance {
 	return Appliance {
 		Name: name,
 		Consumption: consumption,
